requirements.txt
# requirements.txt
pyyaml
pandas
numpy
sqlalchemy
sentence-transformers
scikit-learn
pymilvus
openai
# For other RDBMS support
pymysql
pyodbc
# For future vector DBs
redis
qdrant-client
chromadb
cx_Oracle
streamlit
